+++ Packages/Feedback/ControllerLED/Runtime/IBFControllerLed.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+
+namespace BFTools.Feedback.ControllerLED
+{
+    public interface IBFControllerLed
+    {
+        bool IsSupported { get; }
+        void SetColor(Color color);
+        void TurnOff();
+    }
+}